v.9

+++ build/lib/iBot/files_activities.py
@@ -3,6 +3,8 @@ import shutil
 import time
 import requests
 import PyPDF2
+import PIL
+from shutil import copyfile
 
 
 def convert_bytes(num):
@@ -15,12 +17,16 @@ def convert_bytes(num):
 class File:
 
     def __init__(self, path):
-
-        self.exists = os.path.isfile(path)
         self.path = path
-        self.byteSize = convert_bytes(os.stat(path).st_size)
-        self.creationTime = time.ctime(os.path.getctime(path))
-        self.modificationTime = time.ctime(os.path.getmtime(path))
+        self.exists = os.path.isfile(self.path)
+        if self.exists:
+            self.byteSize = convert_bytes(os.stat(path).st_size)
+            self.creationTime = time.ctime(os.path.getctime(path))
+            self.modificationTime = time.ctime(os.path.getmtime(path))
+        else:
+            self.byteSize = None
+            self.creationTime = None
+            self.modificationTime = None
 
     def rename(self, new_file_name):
         old_path = self.path
@@ -54,59 +60,70 @@ class File:
             elif "/" in self.path:
                 new_path = new_location + "/" + "(" + str(uuid.uuid4())[:8] + ") " + name
             os.rename(self.path, new_path)
-            self.path = new_path
+
+        self.path = new_path
 
     def remove(self):
         if os.path.isfile(self.path):
             os.remove(self.path)
 
-    def copy(self, new_path):
-        from shutil import copyfile
-        copyfile(self.path, new_path)
+    def copy(self, new_location=None):
+        import uuid
+        if new_location is None:
+            if "\\" in self.path:
+                new_location = self.path.replace(self.path.split("\\")[-1], "")
+            elif "/" in self.path:
+                new_location = self.path.replace(self.path.split("/")[-1], "")
 
-    def WaitFor(self):
+        if "\\" in self.path:
+            name = self.path.split("\\")[-1]
+            new_path = new_location + "\\" + name
+        elif "/" in self.path:
+            name = self.path.split("/")[-1]
+            new_path = new_location + "/" + name
+
+        if os.path.exists(self.path):
+            if not os.path.exists(new_path):
+                copyfile(self.path, new_path)
+            elif os.path.exists(new_path):
+                if "\\" in self.path:
+                    new_path = new_location + "\\" + "(" + str(uuid.uuid4())[:8] + ") " + name
+                elif "/" in self.path:
+                    new_path = new_location + "/" + "(" + str(uuid.uuid4())[:8] + ") " + name
+                copyfile(self.path, new_path)
+
+    def waitFor(self):
         from time import sleep
-        while not os.path.exists(self.path):
+        while not os.path.isfile(self.path):
             sleep(1)
-
-    @staticmethod
-    def download(url, path, name=None):
-        if not name:
-            filename = path + "/" + name
-        else:
-            filename = path + "/" + str(url).split('/')[-1]
-        r = requests.get(url, allow_redirects=True)
-        open(filename, 'wb').write(r.content)
-
-        return filename
+        self.__init__(self.path)
 
 
 class PDF(File):
-
     def __init__(self, path):
         File.__init__(self, path)
-
         self.path = path
         self.pages = PyPDF2.PdfFileReader(open(path, "rb")).getNumPages() - 1
         self.info = PyPDF2.PdfFileReader(open(path, "rb")).getDocumentInfo()
         return
 
-    def readPage(self, pageNum):
-        with open(self.path, "rb") as filehandle:
-            pdf = PyPDF2.PdfFileReader(filehandle)
+    def readPage(self, pageNum, encoding=None):
+        if encoding is None:
+            encoding = "utf-8"
+        with open(self.path, "rb") as file:
+            pdf = PyPDF2.PdfFileReader(file)
             page = pdf.getPage(pageNum)
-            text = page.extractText()
+            text = page.extractText().encode(encoding)
+            if text is None:
+                raise ValueError("Pdf not readable with this method, use OCR instead")
             return text
 
     def merge(self, pdf_document2, merged_path):
         from PyPDF2 import PdfFileMerger
-
         pdfs = [str(self.path), str(pdf_document2)]
         merger = PdfFileMerger()
-
         for pdf in pdfs:
             merger.append(pdf)
-
         merger.write(merged_path)
         return
 
@@ -114,46 +131,46 @@ class PDF(File):
 class Image(File):
     def __init__(self, path):
         File.__init__(self, path)
-
-        self.size = str(Image.open(path).size)
-        self.format = Image.open(path).format
+        self.size = self.open().size
+        self.format = self.open().format
 
     def open(self):
-        im = Image.open(self.path)
-        return im.show()
+        im = PIL.Image.open(self.path)
+        return im
 
     def rotate(self, angle):
-        im = Image.open(self.path)
+        im = self.open()
         return im.rotate(angle, expand=True).save(self.path)
 
     def resize(self, size):
-        im = Image.open(self.path)
+        im = self.open()
         return im.resize(size).save(self.path)
 
-    def crop(self, path, box=None):
-        im = Image.open(path)
-        return im.crop(box).save(path)
+    def crop(self, box=None):
+        im = self.open()
+        return im.crop(box).save(self.path)
 
     # Mirrors an image with a given path from left to right.
 
     def mirrorH(self):
-        im = Image.open(self.path)
-        return im.transpose(Image.FLIP_LEFT_RIGHT).save(self.path)
+        im = self.open()
+        return im.transpose(PIL.Image.FLIP_LEFT_RIGHT).save(self.path)
 
     #    Mirrors an image with a given path from top to bottom.
-
     def mirrorV(self):
-        im = Image.open(self.path)
-        return im.transpose(Image.FLIP_TOP_BOTTOM).save(self.path)
+        im = self.open()
+        return im.transpose(PIL.Image.FLIP_TOP_BOTTOM).save(self.path)
 
 
 class Folder:
     def __init__(self, path):
         self.path = path
-        self.name = path.split("\\")[-1:]
-        self.exists = os.path.isdir(path)
-        self.fileList = self.filelist()
-        self.subFoldersList = self.listSubFolders()
+        if not os.path.exists(self.path):
+            os.makedirs(path)
+        if "\\" in path:
+            self.name = path.split("\\")[-1:]
+        elif "/" in path:
+            self.name = path.split("/")[-1:]
 
     def rename(self, new_folder_name):
         old_path = self.path
@@ -202,23 +219,25 @@ class Folder:
                         os.remove(os.path.join(root, name))
                     for name in dirs:
                         os.rmdir(os.path.join(root, name))
-        return
 
-    def copy(self, new_location):
-        old_path = self.path
+    def copy(self, new_location = None):
         import uuid
+        if new_location is None:
+            if "\\" in self.path:
+                new_location = self.path.replace(self.path.split("\\")[-1], "")
+            elif "/" in self.path:
+                new_location = self.path.replace(self.path.split("/")[-1], "")
         if "\\" in new_location:
-            new_path = new_location + "\\" + old_path.split("\\")[-1]
+            new_path = new_location + "\\" + self.path.split("\\")[-1]
         elif "/" in new_location:
-            new_path = new_location + "/" + old_path.split("/")[-1]
-        if os.path.isdir(old_path):
+            new_path = new_location + "/" + self.path.split("/")[-1]
+        if os.path.isdir(self.path):
             if not os.path.isdir(new_path):
-                shutil.copytree(old_path, new_path)
+                shutil.copytree(self.path, new_path)
             elif os.path.isdir(new_path):
                 if os.path.isdir(new_path):
                     new_path = new_path + " (" + str(uuid.uuid4())[:8] + ")"
-                shutil.copytree(old_path, new_path)
-        return
+                shutil.copytree(self.path, new_path)
 
     def listSubFolders(self):
         subfolders = [f.path for f in os.scandir(self.path) if f.is_dir()]
@@ -231,7 +250,11 @@ class Folder:
                 FileList.append(self.path + "/" + File)
         return FileList
 
-    def create(self):
-        if not os.path.exists(self.path):
-            os.makedirs(self.path)
-        return self.path
+    def downloadFile(self, url, name=None):
+        if not name:
+            filename = self.path + "/" + name
+        else:
+            filename = self.path + "/" + str(url).split('/')[-1]
+        r = requests.get(url, allow_redirects=True)
+        open(filename, 'wb').write(r.content)
+        return filename
